fix(vfs): audit blob sync implementation
- Wrap blob sync in withDownloadTracking for UI progress
- Add bigint column type for >2GB blob manifest sizes
- Add v009 migration to widen manifest byte columns
- Add SyncQueueInboundBlobSection tests (5 tests)
- Expand vfsDirectBlobRead tests (6 new tests)
- Add inboundBlobDownloads i18n key to test setup

+++ packages/db/src/schema/definitionCommunicationsVfsBlob.ts
@@ -225,12 +225,12 @@ const vfsBlobManifestsTable: TableDefinition = {
       notNull: true
     },
     totalPlaintextBytes: {
-      type: 'integer',
+      type: 'bigint',
       sqlName: 'total_plaintext_bytes',
       notNull: true
     },
     totalCiphertextBytes: {
-      type: 'integer',
+      type: 'bigint',
       sqlName: 'total_ciphertext_bytes',
       notNull: true
     },
